renamed create_canon_map.c to create-canon-map.c to conform to other naming files
